test/browser/test-actionlistener.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="UTF-8">
  <title>Example tests with testharness.js for Travis CI</title>
  <script src="resources/testharness.js"></script>
  <script src="resources/testharnessreport.js"></script>
  <script src="https://static.robotwebtools.org/EventEmitter2/current/eventemitter2.min.js"></script>
  <script src="https://static.robotwebtools.org/roslibjs/current/roslib.js"></script>
</head>
<body>
  <div id="log"></div>
  <script>
    test(function(t) {
      var ros = new ROSLIB.Ros();
      var example = new ROSLIB.Topic({
        ros: ros,
        name: '/example_topic',
        messageType: 'std_msgs/String'
      });

      assert_true(true, "Nothing wrong happened.");
    }, "Goal sample test");

    async_test(function(t) {
      var ros = new ROSLIB.Ros({url: 'ws://127.0.0.1:9090'});

      var fibonacciServer = new ROSLIB.SimpleActionServer({
        ros: ros,
        serverName: '/fibonacci',
        actionName: 'actionlib_tutorials/FibonacciAction'
      });
      fibonacciServer.on('goal', function(goalMessage) {
        fibonacciSequence = [];
        fibonacciSequence.push(0);
        fibonacciSequence.push(1);

        for (var i = 1; i < goalMessage.order; i++ ) {
          fibonacciSequence.push( fibonacciSequence[i] + fibonacciSequence[i-1] );

        if (that.canceled === true ) {
          that.fibonacciServer.setPreempted();
        }
        //send feedback
        var feedback = { sequence: fibonacciSequence };
          fibonacciServer.sendFeedback(fibonacciSequence);
        }

        //send result
        var result = { sequence: fibonacciSequence};
        fibonacciServer.setSucceeded(result);
      });

      fibonacciServer.on('cancel', t.step_func_done(function() {
        t.done();
      }));
      
      var fibonacciClient = new ROSLIB.ActionClient({
        ros: ros,
        serverName: '/fibonacci',
        actionName: 'actionlib_tutorials/FibonacciAction'
      });

      var goal = new ROSLIB.Goal({
        actionClient: fibonacciClient,
        goalMessage: {
          order: 7
        }
      });
      
      goal.cancel();
    }, 'Goal.cancel() test case 1');

    test(function() {
      var ros = new ROSLIB.Ros({url: 'ws://127.0.0.1:9090'});

      var fibonacciServer = new ROSLIB.SimpleActionServer({
        ros: ros,
        serverName: '/fibonacci',
        actionName: 'actionlib_tutorials/FibonacciAction'
      });
      fibonacciServer.on('goal', function(goalMessage) {
        fibonacciSequence = [];
        fibonacciSequence.push(0);
        fibonacciSequence.push(1);

        for (var i = 1; i < goalMessage.order; i++ ) {
          fibonacciSequence.push( fibonacciSequence[i] + fibonacciSequence[i-1] );

        if (that.canceled === true ) {
          that.fibonacciServer.setPreempted();
        }
        //send feedback
        var feedback = { sequence: fibonacciSequence };
          fibonacciServer.sendFeedback(fibonacciSequence);
        }

        //send result
        var result = { sequence: fibonacciSequence};
        fibonacciServer.setSucceeded(result);
      });
      
      var fibonacciClient = new ROSLIB.ActionClient({
        ros: ros,
        serverName: '/fibonacci',
        actionName: 'actionlib_tutorials/FibonacciAction'
      });

      var goal = new ROSLIB.Goal({
        actionClient: fibonacciClient,
        goalMessage: {
          order: 7
        }
      });
      
      goal.cancel();
      goal.cancel();     
    }, 'Goal.cancel() test case 2');

    async_test(function(t) {
      var ros = new ROSLIB.Ros({url: 'ws://127.0.0.1:9090'});

      var fibonacciServer = new ROSLIB.SimpleActionServer({
        ros: ros,
        serverName: '/listener_fibonacci',
        actionName: 'actionlib_tutorials/FibonacciAction'
      });
      fibonacciServer.on('goal', function(goalMessage) {
        fibonacciSequence = [];
        fibonacciSequence.push(0);
        fibonacciSequence.push(1);

        for (var i = 1; i < goalMessage.order; i++ ) {
          fibonacciSequence.push( fibonacciSequence[i] + fibonacciSequence[i-1] );

        if (that.canceled === true ) {
          that.fibonacciServer.setPreempted();
        }
        //send feedback
        var feedback = { sequence: fibonacciSequence };
          fibonacciServer.sendFeedback(fibonacciSequence);
        }

        //send result
        var result = { sequence: fibonacciSequence};
        fibonacciServer.setSucceeded(result);
      });
      
      var fibonacciClient = new ROSLIB.ActionClient({
        ros: ros,
        serverName: '/listener_fibonacci',
        actionName: 'actionlib_tutorials/FibonacciAction'
      });

      var goal = new ROSLIB.Goal({
        actionClient: fibonacciClient,
        goalMessage: {
          order: 7
        }
      });
      
      var listener = new ROSLIB.ActionListener({
        ros: ros,
        serverName: '/listener_fibonacci',
        actionName: 'actionlib_tutorials/FibonacciAction'
      });
    }, 'ActionListener');    

  </script>
</body>
</html>
